feat: Add NavTab CT styles
